Bring upstream up to date. (#9)
* Add directory iterator that uses readdirplus

* Flip pread back to using a Vec

* Remove libffi

libffi fails to build on centos7 and i'm only using it for
one function that isn't tested yet.  Removing for now.

* Mark Gluster as threadsafe

As far as I can tell the cluster handle is threadsafe.  There's
some bugs that have been opened over the years but they're always
fixed.  I think this should be fine but if bugs crop up I'll
remove it.