Add logging and volfile getter
The gfapi log file can be set now.  Also added a function to get
the volfile.  It's not super useful because it dumps everything
as one giant string.  I plan on changing it in the future to
parse into a proper graph and return that instead.  Adding it
now though just in case anyone really needs it.